pyproject.toml
[tool.poetry]
name = "ist-project"
version = "0.1.0"
description = ""
authors = ["theMossCode <collombuvi@gmail.com>"]
readme = "README.md"
packages = [{include = "ist_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
dj-database-url = "^1.2.0"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
django-guardian = "^2.4.0"
pymongo = "^4.3.3"
bokeh = "^3.0.3"
paho-mqtt = "^1.6.1"
gunicorn = "^20.1.0"
pymysql = "^1.0.2"
channels = "^4.0.0"
daphne = "^4.0.0"
channels-redis = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
